Code forces question
You are given an array a of length n.

Let's define the eversion operation. Let x=an. Then array a is partitioned into two parts: left and right. The left part contains the elements of a that are not greater than x (≤x). The right part contains the elements of a that are strictly greater than x (>x). The order of elements in each part is kept the same as before the operation, i. e. the partition is stable. Then the array is replaced with the concatenation of the left and the right parts.

For example, if the array a is [2,4,1,5,3], the eversion goes like this: [2,4,1,5,3]→[2,1,3],[4,5]→[2,1,3,4,5].

We start with the array a and perform eversions on this array. We can prove that after several eversions the array a stops changing. Output the minimum number k such that the array stops changing after k eversions.