Code forces question
You are given an array a1,a2,…,an of positive integers. A good pair is a pair of indices (i,j) with 1≤i,j≤n such that, for all 1≤k≤n, the following equality holds:

|ai−ak|+|ak−aj|=|ai−aj|,
where |x| denotes the absolute value of x.

Find a good pair. Note that i can be equal to j.