Code forces question
The company "Divan's Sofas" is planning to build n+1 different buildings on a coordinate line so that:

the coordinate of each building is an integer number;
no two buildings stand at the same point.
Let xi be the coordinate of the i-th building. To get from the building i to the building j, Divan spends |xi−xj| minutes, where |y| is the absolute value of y.

All buildings that Divan is going to build can be numbered from 0 to n. The businessman will live in the building 0, the new headquarters of "Divan's Sofas". In the first ten years after construction Divan will visit the i-th building ai times, each time spending 2⋅|x0−xi| minutes for walking.

Divan asks you to choose the coordinates for all n+1 buildings so that over the next ten years the businessman will spend as little time for walking as possible.