Code forces question
You are given a positive integer n. You have to find 4 positive integers a,b,c,d such that

a+b+c+d=n, and
gcd(a,b)=lcm(c,d).
If there are several possible answers you can output any of them. It is possible to show that the answer always exists.

In this problem gcd(a,b) denotes the greatest common divisor of a and b, and lcm(c,d) denotes the least common multiple of c and d.