Code forces Question
m chairs are arranged in a circle sequentially. The chairs are numbered from 0 to m−1. n people want to sit in these chairs. The i-th of them wants at least a[i] empty chairs both on his right and left side.

More formally, if the i-th person sits in the j-th chair, then no one else should sit in the following chairs: (j−a[i])modm, (j−a[i]+1)modm, ... (j+a[i]−1)modm, (j+a[i])modm.

Decide if it is possible to sit down for all of them, under the given limitations.