Code forces QUESTION
You are given an array a of length n. We define the equality of the array as the number of indices 1≤i≤n−1 such that ai=ai+1. We are allowed to do the following operation:

Select two integers i and x such that 1≤i≤n−1 and 1≤x≤109. Then, set ai and ai+1 to be equal to x.
Find the minimum number of operations needed such that the equality of the array is less than or equal to 1.