1616-B
You have a string s1s2…sn and you stand on the left of the string looking right. You want to choose an index k (1≤k≤n) and place a mirror after the k-th letter, so that what you see is s1s2…sksksk−1…s1. What is the lexicographically smallest string you can see?

A string a is lexicographically smaller than a string b if and only if one of the following holds:

a is a prefix of b, but a≠b;
in the first position where a and b differ, the string a has a letter that appears earlier in the alphabet than the corresponding letter in b.