1579-B
The new generation external memory contains an array of integers a[1…n]=[a1,a2,…,an].

This type of memory does not support changing the value of an arbitrary element. Instead, it allows you to cut out any segment of the given array, cyclically shift (rotate) it by any offset and insert it back into the same place.

Technically, each cyclic shift consists of two consecutive actions:

You may select arbitrary indices l and r (1≤l<r≤n) as the boundaries of the segment.
Then you replace the segment a[l…r] with it's cyclic shift to the left by an arbitrary offset d. The concept of a cyclic shift can be also explained by following relations: the sequence [1,4,1,3] is a cyclic shift of the sequence [3,1,4,1] to the left by the offset 1 and the sequence [4,1,3,1] is a cyclic shift of the sequence [3,1,4,1] to the left by the offset 2.
For example, if a=[1,3,2,8,5], then choosing l=2, r=4 and d=2 yields a segment a[2…4]=[3,2,8]. This segment is then shifted by the offset d=2 to the left, and you get a segment [8,3,2] which then takes the place of of the original elements of the segment. In the end you get a=[1,8,3,2,5].

Sort the given array a using no more than n cyclic shifts of any of its segments. Note that you don't need to minimize the number of cyclic shifts. Any method that requires n or less cyclic shifts will be accepted.