XOR Specia-LIS-t
YouKn0wWho has an integer sequence a1,a2,…an. Now he will split the sequence a into one or more consecutive subarrays so that each element of a belongs to exactly one subarray. Let k be the number of resulting subarrays, and h1,h2,…,hk be the lengths of the longest increasing subsequences of corresponding subarrays.

For example, if we split [2,5,3,1,4,3,2,2,5,1] into [2,5,3,1,4], [3,2,2,5], [1], then h=[3,2,1].

YouKn0wWho wonders if it is possible to split the sequence a in such a way that the bitwise XOR of h1,h2,…,hk is equal to 0. You have to tell whether it is possible.

The longest increasing subsequence (LIS) of a sequence b1,b2,…,bm is the longest sequence of valid indices i1,i2,…,ik such that i1<i2<…<ik and bi1<bi2<…<bik. For example, the LIS of [2,5,3,3,5] is [2,3,5], which has length 3.

An array c is a subarray of an array b if c can be obtained from b by deletion of several (possibly, zero or all) elements from the beginning and several (possibly, zero or all) elements from the end.