Stone Age Problem
Once upon a time Mike and Mike decided to come up with an outstanding problem for some stage of ROI (rare olympiad in informatics). One of them came up with a problem prototype but another stole the idea and proposed that problem for another stage of the same olympiad. Since then the first Mike has been waiting for an opportunity to propose the original idea for some other contest... Mike waited until this moment!

You are given an array a of n integers. You are also given q queries of two types:

Replace i-th element in the array with integer x.
Replace each element in the array with integer x.
After performing each query you have to calculate the sum of all elements in the array.